Compute the SP size across multiple cores and tiles.
There was a bug in the computation of the SP size, in which the number of tiles available in the system, were not considered. This commit fixes the issue.

+++ SpikeModel/src/MemoryTile/MemoryCPUWrapper.hpp
@@ -93,7 +93,7 @@ namespace spike_model {
 
 		private:
 			const uint8_t num_of_registers = 32;
-			uint64_t sp_reg_size;
+			size_t sp_reg_size;
 
 			enum class SPStatus {
 				IS_L2 = 0,
